requirements.txt
langchain
langchain-openai
langchain-chroma
openai
gradio
python-dotenv
streamlit
